verificacion inscripcion clase

+++ api/services/user_service.py
@@ -1,6 +1,8 @@
 from models.usuario import Usuario
-from schemas.usuario import UsuarioUpdate
+from repositories.inscripcion_repository import InscripcionRepository
+from repositories.suscripcion_repository import SuscripcionRepository
 from repositories.user_repository import UserRepository
+from schemas.usuario import UsuarioUpdate
 from exceptions.general import NotFoundException, BadRequestException, ConflictException
 from sqlalchemy.ext.asyncio import AsyncSession
 import uuid
@@ -38,6 +40,13 @@ class UserService:
     async def delete(self, usuario_id: uuid.UUID, current_usuario_id: uuid.UUID) -> Usuario:
         if usuario_id == current_usuario_id:
             raise BadRequestException("No podés eliminar tu propio usuario")
+
+        if await InscripcionRepository(self.db).has_active_inscripciones(usuario_id):
+            raise ConflictException("No se puede eliminar al usuario porque está inscripto a una clase.")
+
+        if await SuscripcionRepository(self.db).has_active_suscripciones(usuario_id):
+            raise ConflictException("No se puede eliminar al usuario porque tiene suscripciones activas.")
+
         usuario = await self.repo.delete(usuario_id)
         if not usuario:
             raise NotFoundException("Usuario no encontrado")
@@ -49,6 +58,13 @@ class UserService:
             raise NotFoundException("Usuario no encontrado")
         if usuario.id == current_usuario_id:
             raise BadRequestException("No podés eliminar tu propio usuario")
+
+        if await InscripcionRepository(self.db).has_active_inscripciones(usuario.id):
+            raise ConflictException("No se puede eliminar al usuario porque está inscripto a una clase.")
+
+        if await SuscripcionRepository(self.db).has_active_suscripciones(usuario.id):
+            raise ConflictException("No se puede eliminar al usuario porque tiene suscripciones activas.")
+
         usuario = await self.repo.delete_by_dni(dni)
         if not usuario:
             raise NotFoundException("Usuario no encontrado")
